move renderWithSetup util to test-utils file
