+++ frontend/components/forms/ForgotPasswordForm/ForgotPasswordForm.tests.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 
-import { renderWithSetup } from "test/testingUtils";
+import { renderWithSetup } from "test/test-utils";
 
 import ForgotPasswordForm from "./ForgotPasswordForm";
