+++ frontend/components/forms/RegistrationForm/AdminDetails/AdminDetails.tests.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 
-import { renderWithSetup } from "test/testingUtils";
+import { renderWithSetup } from "test/test-utils";
 
 import AdminDetails from "components/forms/RegistrationForm/AdminDetails";
