+++ frontend/components/forms/RegistrationForm/OrgDetails/OrgDetails.tests.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 
-import { renderWithSetup } from "test/testingUtils";
+import { renderWithSetup } from "test/test-utils";
 import OrgDetails from "components/forms/RegistrationForm/OrgDetails";
 
 describe("OrgDetails - form", () => {
