+++ frontend/components/forms/UserSettingsForm/UserSettingsForm.tests.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { fireEvent, render, screen } from "@testing-library/react";
 
-import { renderWithSetup } from "test/testingUtils";
+import { renderWithSetup } from "test/test-utils";
 
 import UserSettingsForm from "components/forms/UserSettingsForm";
