+++ frontend/test/test-utils.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { render, RenderOptions } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
 
 import { AppContext, IAppContext, initialState } from "context/app";
 
@@ -20,4 +21,16 @@ export const renderWithAppContext = (
     <AppContext.Provider value={value}>{component}</AppContext.Provider>,
     renderOptions
   );
+};
+
+/**
+  A collection of utilities to enable easier writting of tests
+ */
+
+// eslint-disable-next-line import/prefer-default-export
+export const renderWithSetup = (component: JSX.Element) => {
+  return {
+    user: userEvent.setup(),
+    ...render(component),
+  };
 };
